Refine float vs double usage

+++ src/base/processing/LossEstimator.cpp
@@ -38,7 +38,7 @@ void LossEstimator::add_input(FramePtr frame) {
 
 void LossEstimator::process_frame_(const Frame& frame) {
     for (size_t n = 0; n < frame.size(); n++) {
-        auto s = double(frame[n]);
+        auto s = (float)frame[n];
 
         if (signal_runmax_(std::abs(s))
             >= MaxSample * config_.signal_detection_threshold) {
@@ -64,7 +64,7 @@ void LossEstimator::report_losses_() {
     const auto elapsed = limiter_.allow();
 
     if (elapsed > 0) {
-        const double loss_rate = losses_ / elapsed;
+        const double loss_rate = (double)losses_ / elapsed;
         const double avg_loss_rate = sma_(loss_rate);
 
         const double loss_ratio = double(no_signal_) / (signal_ + no_signal_) * 100.0;
